backend/requirements.txt
fastapi==0.68.0
uvicorn==0.14.0
mysql-connector-python